backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
pydantic-settings==2.9.1
google-cloud-storage==3.1.0
